The following changes has been made: 1) Updated documentation on KIM_API_allocate in KIMserviceDescription.txt
2) Updated documentation on KIM_API_get_Units, KIM_API_get_original_Units --
documentation removed (see next item)

3) Removed the functions dealing with units from the KIMserviceDescription.txt file.

4)the bug " The functions KIM_API_get_*_neigh() should not check the number of
neighbors (against the KIM_MAX_NEIGHBORS value) when the Test's routine
returns an error.  (Most likely, in this case the test will not set a value
for the number of neighbors and the returned value will not be initialized.
So, then the value is likely to be larger than the max and result in an
unneeded error message)."
is fixed

5) Changed the name of the function `KIM_API_get_listAtomsTypes' to `KIM_API_get_listAtomTypes'

6) Removed `stressesPerAtom' from the standard.kim file.

7) added new KIM_API_get_neigh_mode() KIMservice routine

and some minor cosmetic changes