bug fix for NBC related special cases
In is_it_match_noFlagCount added code indicate a positive match
if the item is required by one of the NBC's.  This deals with
cases where the Test provides say numberContributingAtoms and
the Model only supports CLUSTER (so doesn't list
numberContrbutingAtoms in its descriptor file).

In do_flag_match added code to handle the case where only CLUSTER
is used and none of the Neigh_*Access lines are provided