Add KIM_API_get_rank and KIM_API_get_rank_by_index functions.
These are useful because KIM_API_get_shape() expects an allocated array
argument with a length of the rank of the parameter. So the only way to
use that was to use an array that was much to big. Now the rank can be
queried and the array allocated as needed.