Fixing the failure when sourcing the environment from zsh.
A minor solution on LINUX and MAC using zsh, esp when installing Kim-API locally.
In zsh, the value of $0 depends on the FUNCTION_ARGZERO option, which is set by default and temporarily sets the $0 to the name of the function or script when sourcing a script. To fix this problem on zsh, we can use POSIX_ARGZERO which exposes the value of $0 in spite of the current FUNCTION_ARGZERO setting.