Changing c_long to c_long_long to avoid overflow.
Since Fortran does not have unsigned integers, we have to use the signed integers. According to the standard, c_long is at least 32 bits in size (no guarantee  of 64 bits in size) and can be different on different platforms.
We use a c_long_long which is at least 64 bits in size, and as long as the signed values are positive and less than the maximum value for an unsigned integer,  we can pass this value to C function.