See text as code block
When using doxygen as documentation engine.
In case in a paragraph a code block is requested, this can be accomplished by means of some indentation relative to the previous text in the paragraph. A problem occurs when there is no previous text (e.g. paragraph 'api') to get the code just insert an empty line in the beginning of such a paragraph.