TST: Add sysuser-shadow to openSUSE zypper installs
This package provides the `useradd` utility, which is used in
Dockerfile.common to create user 'kim'.  It appears that sysuser-shadow
must have previously been a dependency of another zypper package, which
meant it got installed and the openSUSE travis jobs would succeed.
Running the openSUSE job in GHA again revealed that it wasn't getting
installed.