Comment out windows CI for now
This job is getting stuck in an infinite loop both when run on github's
runners and on my local machine.  More specifically, the while loop
inside the 'PrivateReplaceORIGIN' function in
KIM_CollectionsImplementation.cpp is running in an infinite loop when
kim-collections-info.exe is called by `kim-api-collections-management
list`.  Unclear why this problem doesn't occur for linux/mac -- will
create issue soon