TST: Overwrite cmake tarball if present in cache
This is a workaround for github caching bug.  I have observed this step
not being skipped even when a cache hit occurs for the cmake caching
step. This was resulting in the tarball being downloaded as
cmake-3.10.3-Linux-x86_64.tar.gz.1, which tripped up
Dockerfile.ubuntu_cmake310, which does some processing that assumes a
single tarball will be found (in order to prevent having to pass
information about the name of the archive to it).  The simplest thing to
do, then, is to tell wget to overwrite the tarball if it's already
present.